feat: adiciona permissões policy

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserProfile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -104,6 +105,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize('destroy', $user);
         $user->delete();
         return redirect()->route('user.index')->with('success', 'Usuário deletado com sucesso!');
     }
